REF isolate declaration of models
