+++ packages/ng/src/app/date-range-picker/date-range-picker.models.ts
@@ -1,0 +1,11 @@
+import * as moment from 'moment';
+
+export interface DateRangeSelectChoice {
+	range: DateRange;
+	label: string;
+}
+
+export interface DateRange {
+	dateMin: moment.Moment;
+	dateMax: moment.Moment;
+}
