+++ packages/ng/src/index.ts
@@ -1,4 +1,4 @@
 export * from './app/lol/lol.module';
 export * from './app/date-range-picker/date-range-picker.module';
-export {DateRangeSelectChoice, DateRange} from './app/date-range-picker/date-range-picker.component';
+export {DateRangeSelectChoice, DateRange} from './app/date-range-picker/date-range-picker.models';
 export { LuRootModule } from './app/lu-root.module';
